Changed datatype of car ranges (electric range, conventional range, ...) from int to float Changed rounding of remaining[...]Range mathods from
(int) Math.floor(<e.g. fuel*maxRange>)

to

Math.floor(<e.g. fuel*maxRange> *1000.0f) /1000.0f so small rnages are not rounded to 0

Adapted unit tests and replaced expected results in example tests

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/populationsynthesis/serialiser/BevCarFormat.java
@@ -34,7 +34,7 @@ public class BevCarFormat implements SerialiserFormat<BatteryElectricCar> {
 		int capacity = electricCarFormat.capacityOf(data);
 		float initialMileage = electricCarFormat.initialMileageOf(data);
 		float batteryLevel = electricCarFormat.batteryLeveOf(data);
-		int electricRange = electricCarFormat.electricRangeOf(data);
+		float electricRange = electricCarFormat.electricRangeOf(data);
 		float batteryCapacity = electricCarFormat.batteryCapacityOf(data);
 		float minimumChargingLevel = electricCarFormat.minimumChargingLevelOf(data);
 		return Optional.of(new BatteryElectricCar(id, position, segment, capacity, initialMileage, batteryLevel,
